% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iROCK-package.R, R/server.R, R/shiny_options.R,
%   R/ui.R
\name{iROCK}
\alias{iROCK}
\alias{iROCK_server}
\alias{iROCK_options}
\alias{iROCK_ui}
\title{Run the iROCK Shiny application}
\usage{
iROCK(project_dir = getwd(), options = iROCK_options(), ...)

iROCK_server(input, output, session)

iROCK_options()

iROCK_ui()
}
\arguments{
\item{project_dir}{directory where ROCK project files are located.}

\item{options}{the results of \code{\link[=iROCK_options]{iROCK_options()}}}

\item{...}{other parameters passed to \code{\link[shiny:shinyApp]{shiny::shinyApp()}}.}

\item{input}{Shiny input object.}

\item{output}{Shiny output object.}

\item{session}{Shiny session object.}
}
\value{
a list object with configurable options for the iROCK Shiny application, including:
\describe{
\item{fixed_style}{}
\item{color_palette}{}
\item{utterance_highlight_color}{The background color }
}
}
\description{
Run the iROCK Shiny application

iROCK Shiny Server

Configuration options for the iROCK Shiny application.

iROCK Shiny UI
}
