% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\docType{data}
\name{link_rock}
\alias{link_rock}
\title{Link to the ROCK project
Internal function.}
\format{
An object of class \code{shiny.tag} of length 3.
}
\usage{
link_rock
}
\description{
Link to the ROCK project
Internal function.
}
\keyword{datasets}
