% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rock_codebook.R
\name{new_rock_codebook}
\alias{new_rock_codebook}
\alias{new_rock_project}
\alias{update_code}
\alias{new_code}
\title{Create a new ROCK codebook.}
\usage{
new_rock_codebook(yaml_file)

new_rock_project(dir)

update_code(yaml_file, id, ...)

new_code(yaml_file, id, ...)
}
\arguments{
\item{yaml_file}{path to the YAML file (i.e. ROCK_codebook.yml).}

\item{dir}{directory to create the new \verb{_ROCKproject.yml} file.}

\item{id}{the ID for the new code.}

\item{...}{any other fields to include with the new code.}
}
\value{
the YAML file.

the newly parsed codefile (see \code{\link[yaml:read_yaml]{yaml::read_yaml()}})
}
\description{
Create a new ROCK codebook.

Create a new ROCK project.

Edit an existing code

Adds a new code to the ROCK codebook.
}
