% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_utterance.R
\name{update_utterance}
\alias{update_utterance}
\title{Update an utterance within a rock file.}
\usage{
update_utterance(rock, uid, codes, add_code, delete_code)
}
\arguments{
\item{rock}{a parsed ROCK source from either \code{\link[rock:parsing_sources]{rock::parse_source()}} or \code{\link[rock:parsing_sources]{rock::parse_sources()}}.}

\item{uid}{the utterance ID.}

\item{codes}{vector of codes to assign the utterance. Note that specifying this parameter will
ignore the \code{add_code} and \code{delete_code} parameters.}

\item{add_code}{a code to add to the utterance.}

\item{delete_code}{a code to remove from the utterance.}
}
\description{
Update an utterance within a rock file.
}
