% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iROCK-package.R
\name{iROCK}
\alias{iROCK}
\title{Run the iROCK Shiny application}
\usage{
iROCK(...)
}
\arguments{
\item{...}{other parameters passed to \code{\link[shiny:runApp]{shiny::runApp()}}.}
}
\description{
Run the iROCK Shiny application
}
